adding delete urn capabilities
